requirements.txt
# Système de Détection et Identification de Drones
# Dépendances Python

# === Core Scientific Computing ===
numpy>=1.24.0
scipy>=1.10.0

# === SDR / Radio ===
# UHD (USRP Hardware Driver) pour LibreSDR B210mini
# Installation: sudo apt-get install python3-uhd
# Ou via pip si disponible
# uhd>=4.6.0

# === WiFi Capture (pour Remote ID) ===
scapy>=2.5.0
bleak>=0.21.1

# === MQTT ===
paho-mqtt>=1.6.1

# === Configuration ===
PyYAML>=6.0

# === Logging & Utilities ===
python-dateutil>=2.8.2

# === Optional: Visualization (pour debug/analyse) ===
matplotlib>=3.7.0
# Décommenter si besoin de visualiser les spectrogrammes

# === Development & Testing ===
pytest>=7.4.0
pytest-cov>=4.1.0

# === Documentation ===
# sphinx>=7.0.0
# sphinx-rtd-theme>=1.3.0
